Move conversion to multiparameter_conversions.py
