+++ core_tools/GUI/keysight_videomaps/data_saver/quantify.py
@@ -1,125 +1,18 @@
-from typing import List, Any, Optional, Sequence
-
-from functools import partial
 from pathlib import Path
 
 import logging
 
-import numpy as np
-import numpy.typing as npt
-
-from qcodes import Parameter, MultiParameter, Instrument
+from qcodes import Instrument
 
 from quantify_core.measurement import MeasurementControl
 from quantify_core.data.handling import set_datadir, get_datadir, get_latest_tuid
+
+from core_tools.utility.multiparameter_conversions import AwgScanToQuantifyMapper
 
 DEFAULT_DATADIR = Path('./data')
 """The default data directory to use if none is set up for quantify_core"""
 _MEASUREMENT_CONTROL_INSTRUMENT_NAME = 'MC_live_plot_saving'
 """The name to use for the measurement control instance."""
-
-
-class AwgScanToQuantifyMapper:
-
-    INVALID_SUBSTRINGS_IN_PARAM_NAME = {'.'}
-    """Specifies the substrings that cannot occur in the parameter name."""
-
-    def __init__(self, multiparameter: MultiParameter):  # pylint: disable=too-many-locals
-        """Split the MultiParameter that is returned by the pulse_lib function fast_scand1d_qblox() into
-        setter and getter parameters that can be used with Quantify-core MeasurementControl."""
-        # set parameters set the index for get parameters
-        self.set_params: List[Parameter] = []
-        # get parameters return 1 single value
-        self.get_params: List[Parameter] = []
-        self.set_params_setpoints: List[npt.NDArray[Any]] = []
-        self._measurement_data = None
-
-        self._multiparameter = multiparameter
-        # check setpoints are the same for every parameter
-        for array in [
-            multiparameter.setpoint_names,
-            multiparameter.setpoint_labels,
-            multiparameter.setpoint_units,
-            multiparameter.setpoints,
-        ]:
-            self._check_all_equal(array)
-
-        for dim, (name, label, unit, setpoints) in enumerate(
-            zip(
-                multiparameter.setpoint_names[0],  # type: ignore[index]
-                multiparameter.setpoint_labels[0],  # type: ignore[index]
-                multiparameter.setpoint_units[0],  # type: ignore[index]
-                multiparameter.setpoints[0],  # type: ignore[index]
-            )
-        ):
-            old_name = name
-            name = AwgScanToQuantifyMapper._sanitize_param_name(name)
-            logging.debug(f'Building settable Parameter {name} from {old_name}.')
-            # reduce setpoints to 1D
-            np_sp: npt.NDArray[Any] = np.asarray(setpoints)
-            if np_sp.ndim > 1:
-                lower_dims = tuple(range(np_sp.ndim - 1))
-                setpoints_min = np_sp.min(axis=lower_dims)
-                setpoints_max = np_sp.max(axis=lower_dims)
-                if np.all(setpoints_min  != setpoints_max ):
-                    raise Exception(f"Setpoints may only vary on last axis {np_sp}")
-                # use setpoints on 1 axis
-                setpoints_1d = setpoints_min
-            else:
-                setpoints_1d  = np_sp
-            val_map = dict(zip(setpoints_1d, range(len(setpoints_1d))))
-            set_param = Parameter(
-                name=name,
-                label=label,
-                unit=unit,
-                val_mapping=val_map,
-                set_cmd=partial(self._set_index, dim),
-            )
-            self.set_params_setpoints.append(setpoints_1d )
-            self.set_params.append(set_param)
-
-        self._index = [-1] * len(self.set_params)
-        self._start_index = [0] * len(self.set_params)
-
-        for name, label, unit in zip(multiparameter.names, multiparameter.labels, multiparameter.units):
-            old_name = name
-            name = AwgScanToQuantifyMapper._sanitize_param_name(name)
-            logging.debug(f'Building gettable Parameter {name} from {old_name}.')
-
-            get_param = Parameter(
-                name=name,
-                label=label,
-                unit=unit,
-                get_cmd=partial(self._get_value, len(self.get_params)),
-            )
-            self.get_params.append(get_param)
-
-    @classmethod
-    def _sanitize_param_name(cls, name: str) -> str:
-        new_name = name
-        for substring in cls.INVALID_SUBSTRINGS_IN_PARAM_NAME:
-            new_name = new_name.replace(substring, '_')
-        return '_' + new_name
-
-    @staticmethod
-    def _check_all_equal(array: Optional[Sequence[Sequence[Any]]]) -> None:
-        ref_val = array[0]  # type: ignore[index]
-        for value in array[1:]:  # type: ignore[index]
-            if value != ref_val:
-                raise ValueError(f"Parameters not equal: {value} != {ref_val}")
-
-    def _set_index(self, dim: int, value: int) -> None:
-        if self._index[dim] == value:
-            return
-        self._index[dim] = value
-        if self._index == self._start_index:
-            # Starting new fast scan. Force re-running of fast scan at next _get_value()
-            self._measurement_data = None
-
-    def _get_value(self, i: int) -> Any:
-        if self._measurement_data is None:
-            self._measurement_data = self._multiparameter()
-        return self._measurement_data[i][tuple(self._index)]  # type: ignore[index]
 
 
 def save_data(vm_data_parameter, label):
