+++ core_tools/utility/multiparameter_conversions.py
@@ -1,8 +1,14 @@
+from typing import List, Any, Optional
+
 from functools import partial
 from collections.abc import Sequence
-import numpy as np
 
-from qcodes import MultiParameter
+import logging
+
+import numpy as np
+from numpy import typing as npt
+
+from qcodes import MultiParameter, Parameter
 
 
 class ConvertingMultiParameter(MultiParameter):
@@ -152,3 +158,106 @@ class MappedMultiParameter(MultiParameter):
         snapshot = super().snapshot_base(update, params_to_skip_update)
         snapshot['parent'] = self.parameter.snapshot()
         return snapshot
+
+
+class AwgScanToQuantifyMapper:
+
+    INVALID_SUBSTRINGS_IN_PARAM_NAME = {'.'}
+    """Specifies the substrings that cannot occur in the parameter name."""
+
+    def __init__(self, multiparameter: MultiParameter):  # pylint: disable=too-many-locals
+        """Split the MultiParameter that is returned by the pulse_lib function fast_scand1d_qblox() into
+        setter and getter parameters that can be used with Quantify-core MeasurementControl."""
+        # set parameters set the index for get parameters
+        self.set_params: List[Parameter] = []
+        # get parameters return 1 single value
+        self.get_params: List[Parameter] = []
+        self.set_params_setpoints: List[npt.NDArray[Any]] = []
+        self._measurement_data = None
+
+        self._multiparameter = multiparameter
+        # check setpoints are the same for every parameter
+        for array in [
+            multiparameter.setpoint_names,
+            multiparameter.setpoint_labels,
+            multiparameter.setpoint_units,
+            multiparameter.setpoints,
+        ]:
+            self._check_all_equal(array)
+
+        for dim, (name, label, unit, setpoints) in enumerate(
+            zip(
+                multiparameter.setpoint_names[0],  # type: ignore[index]
+                multiparameter.setpoint_labels[0],  # type: ignore[index]
+                multiparameter.setpoint_units[0],  # type: ignore[index]
+                multiparameter.setpoints[0],  # type: ignore[index]
+            )
+        ):
+            old_name = name
+            name = AwgScanToQuantifyMapper._sanitize_param_name(name)
+            logging.debug(f'Building settable Parameter {name} from {old_name}.')
+            # reduce setpoints to 1D
+            np_sp: npt.NDArray[Any] = np.asarray(setpoints)
+            if np_sp.ndim > 1:
+                lower_dims = tuple(range(np_sp.ndim - 1))
+                setpoints_min = np_sp.min(axis=lower_dims)
+                setpoints_max = np_sp.max(axis=lower_dims)
+                if np.all(setpoints_min  != setpoints_max ):
+                    raise Exception(f"Setpoints may only vary on last axis {np_sp}")
+                # use setpoints on 1 axis
+                setpoints_1d = setpoints_min
+            else:
+                setpoints_1d  = np_sp
+            val_map = dict(zip(setpoints_1d, range(len(setpoints_1d))))
+            set_param = Parameter(
+                name=name,
+                label=label,
+                unit=unit,
+                val_mapping=val_map,
+                set_cmd=partial(self._set_index, dim),
+            )
+            self.set_params_setpoints.append(setpoints_1d )
+            self.set_params.append(set_param)
+
+        self._index = [-1] * len(self.set_params)
+        self._start_index = [0] * len(self.set_params)
+
+        for name, label, unit in zip(multiparameter.names, multiparameter.labels, multiparameter.units):
+            old_name = name
+            name = AwgScanToQuantifyMapper._sanitize_param_name(name)
+            logging.debug(f'Building gettable Parameter {name} from {old_name}.')
+
+            get_param = Parameter(
+                name=name,
+                label=label,
+                unit=unit,
+                get_cmd=partial(self._get_value, len(self.get_params)),
+            )
+            self.get_params.append(get_param)
+
+    @classmethod
+    def _sanitize_param_name(cls, name: str) -> str:
+        new_name = name
+        for substring in cls.INVALID_SUBSTRINGS_IN_PARAM_NAME:
+            new_name = new_name.replace(substring, '_')
+        return '_' + new_name
+
+    @staticmethod
+    def _check_all_equal(array: Optional[Sequence[Sequence[Any]]]) -> None:
+        ref_val = array[0]  # type: ignore[index]
+        for value in array[1:]:  # type: ignore[index]
+            if value != ref_val:
+                raise ValueError(f"Parameters not equal: {value} != {ref_val}")
+
+    def _set_index(self, dim: int, value: int) -> None:
+        if self._index[dim] == value:
+            return
+        self._index[dim] = value
+        if self._index == self._start_index:
+            # Starting new fast scan. Force re-running of fast scan at next _get_value()
+            self._measurement_data = None
+
+    def _get_value(self, i: int) -> Any:
+        if self._measurement_data is None:
+            self._measurement_data = self._multiparameter()
+        return self._measurement_data[i][tuple(self._index)]  # type: ignore[index]
